+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added minimal Karma
+- Removed Lodash, jquery and StencilUtil
+- Cache to sessionStorage
+- Added webpack
+- Moved cache to a single method
+- Added callback to updateConfig and onGetPrice
 
 ## [0.0.0] - 2018-07-02
 ### Added